Update to run extension structure
