+++ sample-apps/bundles-price-per-component/extensions/price-per-component-js/src/run.graphql
@@ -1,4 +1,4 @@
-query Input {
+query RunInput {
   presentmentCurrencyRate
   cart {
     lines {
