+++ sample-apps/bundles-price-per-component/extensions/price-per-component-js/src/run.test.js
@@ -1,10 +1,11 @@
 import { describe, it, expect } from "vitest";
-import cartTransform from "./index";
+import { run } from './run';
 
 describe("cart transform function", () => {
   it("returns NO_CHANGES when cart is empty", () => {
     expect(
-      cartTransform({
+      run({
+        presentmentCurrencyRate: 1,
         cart: {
           lines: [],
         },
@@ -14,7 +15,8 @@ describe("cart transform function", () => {
 
   it("returns NO_CHANGES when cart contains no bundles", () => {
     expect(
-      cartTransform({
+      run({
+        presentmentCurrencyRate: 1,
         cart: {
           lines: [
             {
@@ -34,7 +36,8 @@ describe("cart transform function", () => {
 
   it("returns operations when cart contains a bundle", () => {
     expect(
-      cartTransform({
+      run({
+        presentmentCurrencyRate: 1,
         cart: {
           lines: [
             {
@@ -92,7 +95,8 @@ describe("cart transform function", () => {
   describe("error cases", () => {
     it("throws an error when bundledComponentData length is 0", () => {
       expect(() =>
-        cartTransform({
+        run({
+          presentmentCurrencyRate: 1,
           cart: {
             lines: [
               {
