+++ sample-apps/optional-add-ons-cart-transform/extensions/optional-add-ons-js/src/run.graphql
@@ -1,4 +1,4 @@
-query Input {
+query RunInput {
   presentmentCurrencyRate
   cart {
     lines {
